fix: more error handling for files that will never contain docblocks (#321)
Follow on from this PR: https://github.com/readmeio/swagger-inline/pull/196
which added support for ignoring `multiline-extract-comments` from markdown
files. We only accounted for `.md` extensions and not `.markdown`, which is
also valid.

This will fix an error upstream in the readme-micro action which uses
swagger-inline to try and extract openapi operations from code.